Mainchat window
Most visual cruft has been removed, including padding, margin, and so on.
There are buttons to toggle text shadow and time stamps, and to toggle yourself AFK, which will also show a red visual cue to remind you of toggling yourself back.
These buttons and the scrollbar, all on the right edge of the window, won't show unless specifically hovered.
The tabs have been linked to the Frame property, so you can fade them in or our by playing with window opacities.